requirement.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
colorama==0.4.6
Django==5.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
pluggy==1.3.0
pyflakes==3.1.0
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
